Start Menu, Load Attempts, Bugs
-Added basic start menu
-Load feature now works as I would like it to for now
-Load feature now gives user 3 attempts then loads a default character
-Load/SaveAllStates classes added if in case I ever want to implement a
full state save (including map) but because of the way I set up the code
this would take a lot of band-aids and I don't think it is necessary to
do for this project. Basically I'd create map, playermovement, etc
objects inside of encounter and then serialize encounter in the play
class. The calls inside of play would then be routed through encounter
like encounter.map.doStuff() and so on.
-Fixed lava hydra bug